Let get_(un)used_ipaddresses_on_subnet just take a network as argument.
The old dict is no longer needed as they just used one of the elements
in the dict. Also let the functions return lists instead of map objects,
as all consumers did list().

This fixes subnet delete as it just sent the network as argument to
get_used_ipaddresses_on_subnet.