Try to add or remove a PtrOverride when an ipaddress is added or removed.
When adding a new IP, only add a PtrOverride if there is only one more
user of that IP. The PtrOverride will point to the pre-existing host.

When an ipaddress is removed at it has a PtrOverride, just remove it. Can
not guess if any other host should have it.