Use default models.Autofield and plural in related_name.
The default models.Autofield is "id" and no need for a e.g.
cumbersome "ipaddress.ipaddressid". This makes the programming
easier, but might be a bit more ambiguous for the web API users.

Similarly a plural related_field shows that it is actually a set
of object, and not a single one.

This obvisously causes massive API changes.