Swagger requires a serializer_class to be set.
Can be set at class level via "serializer_class", or via
the method "get_serializer_class". We (ab)used "get_queryset"
to set the serializer_class, but as Swagger does not call the
function it can not guess the serializer.

This demonstrates that joining both ForwardZone and ReverseZone
under the same namespace in the "/zones/" API is a bit messy.