Add API for getting host, ipv6 address and mac based on ipv4 address.
For hosts with both an ipv4 and ipv6 address often only the ipv4
address is associated with a mac address. Consequently the dhcp
config based on "/dhcphosts/v6/all" is quite lacking. To help
lazy admins create an API for getting ipv6->mac for hosts which
have an ipv4 address and the ipv6 address is not associated with a
mac.

Also add API to get most recently updated Host and Ipaddress.