Add a unit test to check for auth data in log file
This commit adds a unit test that checks if the username, password
and/or token ended up in the log file as a result of a POST request to
the auth endpoint.

Also, add an upload artifact step to one of the CI pipelines that
uploads the log file from the mreg-cli tests.